requirements.txt
openai==0.28.1
langchain
python-dotenv
tiktoken
chromadb
matplotlib